+++ R/RRPP.support.code.r
@@ -352,7 +352,6 @@ lm.args.from.formula <- function(cl){
   
   if(!is.null(lm.args$data)) {
     lm.args$data <- makeDF(form, lm.args$data, n)
-    lm.args$data$Y <- as.matrix(Y)
   }
   
   if(is.null(lm.args$data)) {
@@ -361,7 +360,11 @@ lm.args.from.formula <- function(cl){
     lm.args$data <- lm.args$data[-1]
   }
   
-  rownames(lm.args$data) <- nms
+  lm.args$data$Y <- Y
+  
+  dfmat <- try(as.matrix(lm.args$data), silent = TRUE)
+  if(!inherits(dfmat, "try-error"))
+    rownames(lm.args$data) <- nms
   
   f <- try(do.call(lm, lm.args), silent = TRUE)
 
@@ -601,8 +604,8 @@ checkers <- function(Terms, exchange, RRPP = TRUE){
   reduced <- exchange$reduced
   full <- exchange$full
   Terms <- exchange$Terms
-  trms <- attr(Terms, "term.labels")
-  k <- length(trms)
+  trms <- names(full)
+  k <- length(full)
   Y <- as.matrix(exchange$model[[1]])
   dims <- dim(Y)
   n <- dims[1]
@@ -865,7 +868,7 @@ anova.parts <- function(exchange, SS){
   full <- exchange$full
   Terms <- exchange$Terms
   trms <- attr(Terms, "term.labels")
-  k <- length(trms)
+  k <- length(full)
   n <- NROW(exchange$model)
   if(k > 0) {
     QRr <- lapply(reduced, function(x) if(!is.null(x$qr)) x$qr else qr(rep(0, n)))
@@ -1492,7 +1495,7 @@ out
 # gets the slopes for groups from a lm.rrpp fit
 # used in pairwise
 getSlopes <- function(fit, x, g){
-  k <- length(fit$LM$term.labels)
+  k <- length(fit$Models$full)
   p <- fit$LM$p
   beta <- fit$LM$random.coef[[k]]
   X <- fit$LM$X
@@ -1526,7 +1529,7 @@ getSlopes <- function(fit, x, g){
 # after constaining covariates to mean values
 # used in pairwise
 getLSmeans <- function(fit, g){
-  k <- length(fit$LM$term.labels)
+  k <- length(fit$Models$full)
   n <- fit$LM$n
   beta <- fit$LM$random.coef[[k]]
   dat <- fit$LM$data
@@ -1748,60 +1751,42 @@ logL <- function(fit, tol = NULL, pc.no = NULL){
   gls <- fit$LM$gls
   w <- fit$LM$weights
   Pcov <- fit$LM$Pcov
+  Res <- if(gls) fit$LM$gls.residuals else fit$LM$residuals
   
   if(gls){
-    if(!is.null(Pcov)) Sig <- crossprod(Pcov %*% fit$LM$gls.residuals)/n else
-      Sig <- crossprod(fit$LM$gls.residuals * sqrt(w))/n
+    if(!is.null(Pcov)) Sig <- crossprod(Pcov %*% Res)/n else
+      Sig <- crossprod(Res * sqrt(w))/n
     
   }  else {
     
-    Sig <- crossprod(fit$LM$residuals) /n
+    Sig <- crossprod(Res) /n
     
   }
   
   s <- svd(Sig, nu = 0, nv = k)
   sdev <- s$d/sqrt(max(1, n - 1))
   rank <- min(sum(sdev > (sdev[1L] * tol)), k)
-  if(rank < k){
-    pr <- seq_len(rank)
-    s$v <- s$v[, pr, drop = FALSE]
-  }
-  
-  Yc <- if(fit$LM$gls) scale(Y, scale = F, center = fit$LM$gls.mean) else
-    center(Y)
-  
-  if(p > 1) P <- as.matrix(Yc %*% s$v) else P <- Yc
-  
-  fit$LM$data$Y <- P
-  form <- formula(fit$LM$Terms)
-  form <- update(form, Y ~ .)
-  
-  if(!is.null(w)) {
-    pfit <- lm.rrpp(form, print.progress = FALSE, 
-                    Cov = fit$LM$Cov, data = fit$LM$data, 
-                    iter = 0, weights = w )
-  } else {
-    pfit <- lm.rrpp(form, print.progress = FALSE, 
-                    Cov = fit$LM$Cov, data = fit$LM$data, 
-                    iter = 0)
-  }
-  
+  pr <- pr <- seq_len(rank)
+  s$v <- s$v[, pr, drop = FALSE]
+
+  P <- Res %*% s$v
+
   if(gls){
-    if(!is.null(Pcov)) Sig <- crossprod(Pcov %*% pfit$LM$gls.residuals)/n else
-      Sig <- crossprod(pfit$LM$gls.residuals * sqrt(w))/n
-    if(kappa(Sig) > 1e10) Sig <- RiReg(Sig, pfit$LM$gls.residuals)
+    if(!is.null(Pcov)) Sig <- crossprod(Pcov %*% P)/n else
+      Sig <- crossprod(P * sqrt(w))/n
+    if(kappa(Sig) > 1e10) Sig <- RiReg(Sig, P)
     
   }  else {
     
-    Sig <- crossprod(pfit$LM$residuals) /n
-    if(kappa(Sig) > 1e10) Sig <- RiReg(Sig, pfit$LM$residuals)
+    Sig <- crossprod(P) /n
+    if(kappa(Sig) > 1e10) Sig <- RiReg(Sig, P)
     
   }
   
   if(gls) {
     if(!is.null(Pcov)) {
       ll <- -0.5*(n * rank + n * determinant(Sig, logarithm = TRUE)$modulus[1] + 
-                    rank * determinant(pfit$LM$Cov, logarithm = TRUE)$modulus[1] + 
+                    rank * determinant(fit$LM$Cov, logarithm = TRUE)$modulus[1] + 
                     n * rank * log(2*pi))
     } else {
       ll <- -0.5*(n * rank + n * determinant(Sig, logarithm = TRUE)$modulus[1] -