+++ NEWS.md
@@ -1,5 +1,11 @@
+# CHANGES IN RRPP 1.1.1
 
-# CHANGES IN RRPP 1.0.0.9000
+### BUG FIXES
+* Fixed issue in `lm.args.from.formula` for intercept only models with covariance matrices.
+* Fixed issue with non-full rank design matrices and dropped terms in `lm.rrpp` support functions.
+* Fixed issue in `logL` support function for non-full rank design matrices.
+
+# CHANGES IN RRPP 1.1.0
 
 ### NEW FEATURES
 * `na.omit.rrpp.data.frame` added for handling missing data.